--- apps/ai-dial-admin/src/utils/formatting/boolean.ts
@@ -1,5 +1,0 @@
-import { BasicI18nKey } from '@/src/constants/i18n';
-
-export const formatRequired = (value: string, t: (stringToTranslate: string) => string) => {
-  return value ? t(BasicI18nKey.Yes) : t(BasicI18nKey.No);
-};
